% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_not_PH}
\alias{data_not_PH}
\title{A simulation with three groups without the Proportional Hazards hypothesis}
\format{
\subsection{\code{data_not_PH}}{

A data frame with 600 rows and 3 columns:
\describe{
\item{time}{Time of events in months with truncation at 60 months}
\item{status}{Indicator of censorship. 1 denotes an event, 0 denotes a censor}
\item{arm}{Integer from 0 to 2. Indicates the group the patient belongs to}
}
}
}
\usage{
data_not_PH
}
\description{
This dataframe represents a simulation of a study where the three survival curves
cross each other.
}
\keyword{datasets}
