% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCCT-package.R
\docType{package}
\name{MSCCT-package}
\alias{MSCCT}
\alias{MSCCT-package}
\title{MSCCT: Multiple Survival Crossing Curves Tests}
\description{
Comparison of two or more survival curves whether the proportional hazards hypothesis is verified or not.
}
\author{
\strong{Maintainer}: Hugo MINA PASSI \email{Hugo_Mina-Passi@etu.u-bourgogne.fr}

}
\keyword{internal}
