% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_lr.R
\name{multi_lr}
\alias{multi_lr}
\title{(Weighted) Log-rank test for comparison of two or more survival curves.}
\usage{
multi_lr(df, weights = numeric(), test = "lr", rho = 1, gamma = 0)
}
\arguments{
\item{df}{A dataframe with columns :
\itemize{
\item \code{time} : positive numbers, time-to-event;
\item \code{status} : integer of factor. 0 is (right) censoring, 1 is event;
\item \code{arm} : A vector or something that can be transformed into one.
The group the patient belongs to.
}}

\item{weights}{An object that can be transformed into a matrix. The weights
used for the tests. Default is vector of ones, corresponding to
the usual log-rank test (see Details);}

\item{test}{If weights is not given, specifies the test to perform. Possible values are
\code{lr} for log-rank, \code{gw} for Gehan-Wilcoxon, and \code{fh} for Flemming-Harrington;}

\item{rho, gamma}{The parameters for Flemming-Harrington test. Default is (rho,gamma)=(1,0),
which is also called the Peto-Peto test.}
}
\value{
An object of class \code{multi_lr} containing:
\itemize{
\item \code{U} : Statistics of tests;
\item \code{p} : The corresponding p-values;
\item \code{df} : Degrees of freedom of the statistics of tests;
\item \code{test} : The performed test.
}
}
\description{
Performs a global log-rank test for comparing two or more survival curves.
}
\details{
For the log-rank test, all the weights are equal to one. Gehan-Wilcoxon uses the
number of patients still at risk as weights. For the Flemming-Harrington test,
the weights are S(t)^rho * (1-S(t))^gamma where S is the global survival function.

If \code{weights} is a vector, performs one log-rank test with the given weights.
Weights are given in the same order as time of event : the first weight correspond
to the earliest time of event, and so on. Throws an error if the number of given
weights is not equal to the number of distinct time of events. Returns the test
statistic and the p-value.

If \code{weights} is a matrix (or a two-dimensional object), each columns correspond to
a set of weights. Throws an error if the number of \code{nrow(weights)} is not equal to
the number of distinct time of events. Returns \code{ncol(weights)} test statistics and
the corresponding p-value.
}
\examples{
# Log-rank test
multi_lr(data_not_PH)

# Gehan-Wilcoxon test
multi_lr(data_not_PH)

# It is possible to run several tests with different weights at a time
evt_time = unique(data_not_PH$time[data_not_PH$status == 1])
nb_evt_time = length(evt_time)
weights = matrix(runif(nb_evt_time*3), ncol=3)
multi_lr(data_not_PH, weights=weights)
}
