% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ts.R
\name{multi_ts}
\alias{multi_ts}
\title{Two-staged test for comparison of two or more survival curves.}
\usage{
multi_ts(df, eps = 0.1, nboot = 100, method = p.adjust.methods)
}
\arguments{
\item{df}{A dataframe with columns :
\itemize{
\item \code{time} : positive numbers, time-to-event;
\item \code{status} : integer of factor. 0 is (right) censoring, 1 is event;
\item \code{arm} : integers from 0 to n-1 or factor with at least 2 levels.
The group the patient belongs to.
}}

\item{eps}{A number from 0 to 0.5. See reference for interpretation;}

\item{nboot}{A positive integer, number of bootstrap sample for the second stage;}

\item{method}{The correction used for the p-values. Must be in \link{p.adjust.methods}. Unused for exactly two groups;}
}
\value{
An object of class \code{multi_ts} containing:
\itemize{
\item \code{results} : A matrix. Each row represents a comparison of two curves and contains
the p-values for both stage, the global p-value and the adjusted global p-value;
\item \code{p} : The global p-value for the global test;
\item \code{nb_tests} : The number of performed Two-Stage tests;
\item the parameters \code{eps}, \code{method} and \code{nboot}.
}
}
\description{
Performs a Two-Stage test for each pair of survival curves and apply a correction
in case of several comparisons.
}
\details{
The first stage is a log-rank test. If the first test is significant, then the whole
procedure stops and we conclude that the survival curves are different. If it is not
significant, then the survival curves are either equal or crossing each other and
the log-rank test can't conclude the difference. A second test is performed to distinguish
these two cases.
}
\examples{
multi_ts(data_not_PH, eps=0.05, nboot=100, method="BH")
}
\references{
\itemize{
\item Qiu, P., & Sheng, J. (2008). A two-stage procedure for comparing
hazard rate functions. Journal of the Royal Statistical Society Series
B: Statistical Methodology, 70(1), 191-208. Chen, Zhongxue & Huang, Hanwen & Qiu, Peihua. (2017).
\item An improved two-stage procedure to compare hazard curves. Journal of Statistical
Computation and Simulation. 87. 1-10. 10.1080/00949655.2017.1292276.
}
}
