% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_lr.R
\name{print.multi_lr}
\alias{print.multi_lr}
\title{Print method for the multiple log-rank test}
\usage{
\method{print}{multi_lr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multi_lr} as returned by \code{\link[=multi_lr]{multi_lr()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\value{
None
}
\description{
Print method for the multiple log-rank test
}
\examples{
  x = multi_lr(data_not_PH)
  print(x)


}
