% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rmst.R
\name{print.multi_rmst}
\alias{print.multi_rmst}
\title{Print method for the multiple test of RMST}
\usage{
\method{print}{multi_rmst}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multi_rmst} as returned by \code{\link[=multi_rmst]{multi_rmst()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\description{
Print method for the multiple test of RMST
}
