% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ts.R
\name{print.multi_ts}
\alias{print.multi_ts}
\title{Print method for the multiple Two-Stage test}
\usage{
\method{print}{multi_ts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multi_ts} as returned by \code{\link[=multi_ts]{multi_ts()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\description{
Print method for the multiple Two-Stage test
}
