% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLR.R
\name{print.multiLR}
\alias{print.multiLR}
\title{Print method for the multiple log-rank test}
\usage{
\method{print}{multiLR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multiLR} as returned by \code{\link[=multiLR]{multiLR()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\description{
Print method for the multiple log-rank test
}
