% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirmst.R
\name{print.multirmst}
\alias{print.multirmst}
\title{Print method for the multiple test of RMST}
\usage{
\method{print}{multirmst}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multirmst} as returned by \code{\link[=multirmst]{multirmst()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\description{
Print method for the multiple test of RMST
}
