% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTS.R
\name{print.multiTS}
\alias{print.multiTS}
\title{Print method for the multiple Two-Stage test}
\usage{
\method{print}{multiTS}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multiTS} as returned by \code{\link[=multiTS]{multiTS()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\description{
Print method for the multiple Two-Stage test
}
