% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ts.R
\name{print.multi_ts}
\alias{print.multi_ts}
\title{Print method for the multiple Two-Stage test}
\usage{
\method{print}{multi_ts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multi_ts} as returned by \code{\link[=multi_ts]{multi_ts()}};}

\item{...}{For compatibility with the \code{print} method, unused and to be ignored.}
}
\value{
None
}
\description{
Print method for the multiple Two-Stage test
}
\examples{
  # test with a quarter of the data frame data_not_PH
  ind = c(1:100, 401:500, 801:900)
  x = multi_ts(data_not_PH[ind,], method = "BH", eps = 0.1, nboot = 10)
  print(x)

}
